lexer rewrite done

+++ rs/src/objects/errors.rs
@@ -431,7 +431,7 @@ impl ZyxtError {
                                     .into_iter()
                                     .enumerate()
                                     .filter(|(i, _)| {
-                                        pos.line - 3 <= *i as u32 && *i as u32 <= end_pos.line + 1
+                                        pos.line as isize - 3 <= *i as isize && *i as u32 <= end_pos.line + 1
                                     })
                                     .map(|(_, s)| format!("  {}", s))
                                     .collect::<Vec<_>>()
